use CLOCK_MONOTONIC_RAW if available
Wine's audio drivers timestamp MIDI messages with CLOCK_MONOTONIC_RAW
when available. PipeWire's graph only uses CLOCK_MONOTONIC. Leads to
Aroblems in Ableton Live. Adds a little bit of overhead, and increases
the violation of ASIO's contract that `nsec` should be the the timestamp
of the current sample and not of current processing time, but should be
fine.