separate driver and host priorities
also ditch module-rt and use the thread we created for everything. might
require way higher priority, hence the separation. also just found out
that pw_stream_trigger_process wasn't doing anything since the stream
can't be a driver if connected to ALSA hardware. the fact that buffers 0
and 1 aren't being flipped is currently depending on a race condition
and i've seen it fail on occasion. might have to roll my own spa node to
get this to behave consistently.